+++ .eslintrc
@@ -1,7 +1,8 @@
 {
     "extends": "airbnb-base",
     "rules": {
-      "no-underscore-dangle": ["error", { "allow": ["_id"] }]
+      "no-underscore-dangle": ["error", { "allow": ["_id"] }],
+      "no-console": "off"
     }
 
 }
