+++ config/rateLimiter/rateLimiter.js
@@ -1,0 +1,9 @@
+const rateLimit = require('express-rate-limit');
+
+
+const apiLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 15,
+});
+
+module.exports = { apiLimiter };
