+++ routes/index.js
@@ -1,8 +1,6 @@
 const router = require('express').Router();
 
-const { celebrate, Joi } = require('celebrate');
-
-const rateLimit = require('express-rate-limit');
+const { celebrate } = require('celebrate');
 
 const { createUser, signin } = require('../controllers/controllerUser');
 
@@ -15,27 +13,13 @@ const routerArticle = require('./routerArticle');
 const NotFoundError = require('../errors/not_found_error');
 
 const { messegNotFoundError } = require('../variables/variables');
+const { apiLimiter } = require('../config/rateLimiter/rateLimiter');
+const { celebrateSignin, celebrateSignup } = require('../config/celebrate/celebrate');
 
 
-const apiLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-});
+router.post('/signin', apiLimiter, celebrate(celebrateSignin), signin);
 
-router.post('/signin', apiLimiter, celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().regex(/^([\w0-9_-]\.)*[\w0-9_-]+@[\w0-9_-]+(\.[\w0-9_-]+)*\.\w{2,6}$/),
-    password: Joi.string().required().min(6),
-  }),
-}), signin);
-
-router.post('/signup', apiLimiter, celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().regex(/^([\w0-9_-]\.)*[\w0-9_-]+@[\w0-9_-]+(\.[\w0-9_-]+)*\.\w{2,6}$/),
-    password: Joi.string().required().min(6),
-    name: Joi.string().required().min(2).max(30),
-  }),
-}), createUser);
+router.post('/signup', apiLimiter, celebrate(celebrateSignup), createUser);
 
 router.use(auth);
