+++ variables/variables.js
@@ -1,10 +1,9 @@
-const messegForbidden =  'это статья не Выша, её нельзя удалить';
+const messegForbidden = 'это статья не Выша, её нельзя удалить';
 const messegNotFoundErrorID = 'не правильно указан ID';
 const messegNotFoundError = 'Запрашиваемый ресурс не найден';
 const messegUnauthorized = 'Необходима авторизация!!!';
 
 
-
-
-
-module.exports = { messegForbidden, messegNotFoundErrorID, messegUnauthorized, messegNotFoundError };
+module.exports = {
+  messegForbidden, messegNotFoundErrorID, messegUnauthorized, messegNotFoundError,
+};
